feat(components): rearrange components markup as required by the task
