--- src/components/FilmsList.js
@@ -1,16 +1,0 @@
-import {getFilmCard} from './FilmCard';
-
-export const getFilmsList = (films) => {
-  const filmElements = films.map((film) => getFilmCard(film));
-
-  return `
-  <section class="films-list">
-    <h2 class="films-list__title visually-hidden">All movies. Upcoming</h2>
-
-    <div class="films-list__container">
-      ${filmElements.join(``)}
-    </div>
-
-    <button class="films-list__show-more">Show more</button>
-  </section>`;
-};
