+++ src/components/Profile.js
@@ -2,7 +2,7 @@
  * Returns Header Profile element markup
  * @return {string} element markup
  */
-export const getHeaderProfile = () => `
+export const getProfile = () => `
 <section class="header__profile profile">
   <p class="profile__rating">Movie Buff</p>
   <img class="profile__avatar" src="images/bitmap@2x.png" alt="Avatar" width="35" height="35">
