+++ src/components/ShowMore.js
@@ -1,0 +1,8 @@
+/**
+ * Returns Show More button element markup
+ * @return {string} element markup
+ */
+export const getShowMore = () => `
+  <button class="films-list__show-more">
+    Show more
+  </button>`;
